Add "deleted" flag to Item.
To make space for this the item type shrinks from 32bit to 16bit
which is plenty.

+++ include/osmium/osm/item_type.hpp
@@ -37,7 +37,7 @@ DEALINGS IN THE SOFTWARE.
 
 namespace osmium {
 
-    enum class item_type : uint32_t {
+    enum class item_type : uint16_t {
 
         undefined                              = 0x00,
         node                                   = 0x01,
